Lambda/requirements.txt
mysql-connector-python==8.0.26
bcrypt==3.2.0